Basic skelatal set up for the events page
Created simplified event card for later styling
